Add files via upload
In this lab, i create and change folder permissions using the Windows Command Line Interface (CLI), known as Powershell. In this exercise, you'll:

access administrative privileges to use Powershell in Windows.
view file and folder permissions using the GUI and Powershell commands.
modify the permissions for both files and directories by granting and removing specific permissions using ICACLS in Powershell.
modify the permissions for groups using the GUI and Powershell.